examples/ray-automl-example/automl_service/cloud_native/python/requirements.txt
aiogrpc
statsforecast
scikit-learn
pickleshare
numpy
protobuf
requests